refactor: 💡 Change the structure of services
Changed the structure so that certain calls can be made on more than one
page without having to rewrite the code. The isOpen prop has been
removed from the card component, this verification will be done within
the component. The same happened in the restaurant interface, initially
i thought the return would como from the API.